Added privileges check and fixed routes
Priviliges check was added to only show users links that they are allowed to access. This is already working in most of the side menu links but some links doesn't have it yet because I'm not sure what privileges are required for them.

Spinner is added to login page so that the user has an indication that the page is processing their request.

Rearranged the views and components: divided them between /admin and /student folders.  Also modified router.js to merge admin and student(online registration) routes.  All student routes can be accessed from the root / while admin routes can be accessed from  /admin